+++ examples/dora/image-utils/Cargo.toml
@@ -1,12 +1,10 @@
 [package]
-name = "dora-video-capture"
+name = "dora-image-utils"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
 arrow = "54.2.1"
 dora-node-api = { version = "0.3", features = ["tracing"] }
-dora-image-utils = { path = "../image-utils" }
 eyre = "0.6"
-rand = "0.9"
-kornia = { version = "0.1.8", features = ["gstreamer"] }
+kornia = { version = "0.1.8"}
