+++ examples/dora/imgproc/src/main.rs
@@ -1,10 +1,6 @@
-use dora_node_api::{
-    self, dora_core::config::DataId, ArrowData, DoraNode, Event, IntoArrow, Metadata, Parameter,
-};
-use kornia::{
-    image::{Image, ImageSize},
-    imgproc,
-};
+use dora_image_utils::{arrow_to_image, image_to_arrow};
+use dora_node_api::{self, dora_core::config::DataId, DoraNode, Event};
+use kornia::{image::Image, imgproc};
 
 fn main() -> eyre::Result<()> {
     let (mut node, mut events) = DoraNode::init_from_env()?;
@@ -15,26 +11,21 @@ fn main() -> eyre::Result<()> {
         match event {
             Event::Input { id, metadata, data } => match id.as_str() {
                 "frame" => {
-                    // deserialize the frame
-                    let img = deserialize_frame(data, metadata.clone())?;
+                    // convert the frame to an image
+                    let img = arrow_to_image(data, metadata.clone())?;
 
                     // compute the sobel edge map
                     let mut out = Image::from_size_val(img.size(), 0f32)?;
                     imgproc::filter::sobel(&img.cast()?, &mut out, 3)?;
 
-                    let out_u8 = out.map(|x| *x as u8);
+                    // TODO: make this more efficient in kornia-image crate
+                    let out_u8 = {
+                        let x = out.map(|x| *x as u8);
+                        Image::new(out.size(), x.into_vec())?
+                    };
 
-                    let mut meta_parameters = metadata.parameters;
-                    meta_parameters
-                        .insert("cols".to_string(), Parameter::Integer(img.cols() as i64));
-                    meta_parameters
-                        .insert("rows".to_string(), Parameter::Integer(img.rows() as i64));
-
-                    node.send_output(
-                        output.clone(),
-                        meta_parameters,
-                        out_u8.as_slice().to_vec().into_arrow(),
-                    )?;
+                    let (meta_parameters, data) = image_to_arrow(out_u8, metadata)?;
+                    node.send_output(output.clone(), meta_parameters, data)?;
                 }
                 other => eprintln!("Ignoring unexpected input `{other}`"),
             },
@@ -49,27 +40,4 @@ fn main() -> eyre::Result<()> {
     }
 
     Ok(())
-}
-
-fn deserialize_frame(data: ArrowData, metadata: Metadata) -> eyre::Result<Image<u8, 3>> {
-    let img_cols = metadata.parameters.get("cols").unwrap();
-    let img_cols: i64 = match img_cols {
-        Parameter::Integer(i) => *i,
-        _ => return Err(eyre::eyre!("cols is not an integer")),
-    };
-    let img_rows = metadata.parameters.get("rows").unwrap();
-    let img_rows: i64 = match img_rows {
-        Parameter::Integer(i) => *i,
-        _ => return Err(eyre::eyre!("rows is not an integer")),
-    };
-
-    let img_data: Vec<u8> = TryFrom::try_from(&data)?;
-
-    Ok(Image::new(
-        ImageSize {
-            width: img_cols as usize,
-            height: img_rows as usize,
-        },
-        img_data,
-    )?)
 }
