+++ examples/dora/video-capture/src/main.rs
@@ -1,4 +1,5 @@
-use dora_node_api::{self, dora_core::config::DataId, DoraNode, Event, IntoArrow, Parameter};
+use dora_image_utils::image_to_arrow;
+use dora_node_api::{self, dora_core::config::DataId, DoraNode, Event};
 use kornia::io::stream::{RTSPCameraConfig, V4L2CameraConfig};
 
 fn main() -> eyre::Result<()> {
@@ -42,18 +43,9 @@ fn main() -> eyre::Result<()> {
                         continue;
                     };
 
-                    let mut meta_parameters = metadata.parameters;
-                    meta_parameters
-                        .insert("cols".to_string(), Parameter::Integer(frame.cols() as i64));
-                    meta_parameters
-                        .insert("rows".to_string(), Parameter::Integer(frame.rows() as i64));
+                    let (meta_parameters, data) = image_to_arrow(frame, metadata)?;
 
-                    node.send_output(
-                        output.clone(),
-                        meta_parameters,
-                        // TODO: avoid to_vec copy
-                        frame.as_slice().to_vec().into_arrow(),
-                    )?;
+                    node.send_output(output.clone(), meta_parameters, data)?;
                 }
                 other => eprintln!("Ignoring unexpected input `{other}`"),
             },
